+++ src/main/java/com/telepathicgrunt/the_bumblezone/client/BumblezoneClient.java
@@ -50,6 +50,7 @@ public class BumblezoneClient {
         forgeBus.addListener(FluidClientOverlay::sugarWaterFluidOverlay);
         forgeBus.addListener(FluidClientOverlay::renderHoneyFog);
         forgeBus.addListener(PileOfPollenRenderer::pileOfPollenOverlay);
+        forgeBus.addListener(StingerSpearCooldownClient::isOnCooldown);
     }
 
     public static void onClientSetup(FMLClientSetupEvent event) {