+++ src/main/java/com/telepathicgrunt/the_bumblezone/mixin/items/HoneyShieldDisableMixin.java
@@ -1,11 +1,7 @@
 package com.telepathicgrunt.the_bumblezone.mixin.items;
 
-import com.telepathicgrunt.the_bumblezone.items.HoneyCrystalShieldBehavior;
 import com.telepathicgrunt.the_bumblezone.modinit.BzItems;
-import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.AxeItem;
-import net.minecraft.world.item.Items;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;